feat: prevent ilegal command

+++ app/commands/MigrationCommand.php
@@ -11,6 +11,7 @@ use System\Support\Facades\Schema;
 use function System\Console\fail;
 use function System\Console\info;
 use function System\Console\ok;
+use function System\Console\option;
 use function System\Console\style;
 use function System\Console\warn;
 
@@ -53,9 +54,12 @@ class MigrationCommand extends Command
           ],
           'options'   => [
             '--dry-run' => 'Excute migration but olny get query  output.',
+            '--force'   => 'Force runing migration/databe query in production',
           ],
           'relation'  => [
-            'migrate'  => ['[migration_name]', '--dry-run'],
+            'migrate'          => ['[migration_name]', '--dry-run', '--force'],
+            'database:create'  => ['--force'],
+            'database:drop'    => ['--force'],
           ],
         ];
     }
@@ -65,8 +69,23 @@ class MigrationCommand extends Command
         return app()->get('dsn.sql')['database_name'];
     }
 
+    private function runInDev(): bool
+    {
+        if (app()->isDev() || $this->force) {
+            return true;
+        }
+
+        return option(style('Runing migration/datbase in production?')->textRed(), [
+            'yes' => fn () => true,
+            'no'  => fn () => false,
+        ]);
+    }
+
     public function main()
     {
+        if (false === $this->runInDev()) {
+            return;
+        }
         $print   = new Style("\n");
         $dir     = base_path('/database/migration/');
         $migrate = [];
@@ -107,9 +126,24 @@ class MigrationCommand extends Command
         $print->out(false);
     }
 
-    public function databaseCreate(): void
+    public function databaseCreate(): int
     {
         $db_name = $this->DbName();
+
+        if (!$this->runInDev()) {
+            return 2;
+        }
+
+        /** @var bool */
+        $continue = option(style("Do you want to create database `{$db_name}`?")->textBlue(), [
+            'yes' => fn () => true,
+            'no'  => fn () => false,
+        ]);
+
+        if (false === $continue) {
+            return 2;
+        }
+
         info("creating database `{$db_name}`")->out(false);
 
         $success = Schema::create()->database($db_name)->ifNotExists()->execute();
@@ -117,15 +151,32 @@ class MigrationCommand extends Command
         if ($success) {
             ok("success create database `{$db_name}`")->out(false);
 
-            return;
+            return 0;
         }
 
         fail("cant created database `{$db_name}`")->out(false);
+
+        return 1;
     }
 
-    public function databaseDrop()
+    public function databaseDrop(): int
     {
+        if (!$this->runInDev()) {
+            return 2;
+        }
+
         $db_name = $this->DbName();
+
+        /** @var bool */
+        $continue = option(style("Do you want to drop database `{$db_name}`")->textRed(), [
+            'yes' => fn () => true,
+            'no'  => fn () => false,
+        ]);
+
+        if (false === $continue) {
+            return 2;
+        }
+
         info("try to drop database `{$db_name}`")->out(false);
 
         $success = Schema::drop()->database($db_name)->ifExists(true)->execute();
@@ -133,10 +184,12 @@ class MigrationCommand extends Command
         if ($success) {
             ok("success drop database `{$db_name}`")->out(false);
 
-            return;
+            return 0;
         }
 
         fail("cant drop database `{$db_name}`")->out(false);
+
+        return 1;
     }
 
     public function databaseShow()
